feat: EnableEdgeToEdge and add the necessary padding for the SystemBar to fix Android 35 and above UI. The app uses the MaterialToolbar and doesn't automatically add the required padding.
Also, decide the status of StatusBar and SystemBar Style Status, Light or Dark, based on the background color applied in the code.
The StatusBar and SystemBar should match the app background or be transparent to fit the new material design theme.
Issue: The StatusBar background currently overlap with SystemBar in Landscape mode.
More detail:
https://developer.android.com/develop/ui/views/layout/edge-to-edge